src/requirements.txt
pydantic==1.10.8
pymongo==4.5.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.24.0.post1
pandas==2.0.3
fastapi
joblib==1.4.2
numpy==1.25.2
keras==2.15.0
tensorflow==2.15.0
pillow==9.4.0